Sylius Stable Release

+++ src/Sylius/Component/Core/Uploader/ImageUploader.php
@@ -9,6 +9,8 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace Sylius\Component\Core\Uploader;
 
 use Gaufrette\Filesystem;
@@ -16,25 +18,34 @@ use Sylius\Component\Core\Model\ImageInterface;
 
 class ImageUploader implements ImageUploaderInterface
 {
+    /**
+     * @var Filesystem
+     */
     protected $filesystem;
 
+    /**
+     * @param Filesystem $filesystem
+     */
     public function __construct(Filesystem $filesystem)
     {
         $this->filesystem = $filesystem;
     }
 
-    public function upload(ImageInterface $image)
+    /**
+     * {@inheritdoc}
+     */
+    public function upload(ImageInterface $image): void
     {
         if (!$image->hasFile()) {
             return;
         }
 
-        if (null !== $image->getPath()) {
+        if (null !== $image->getPath() && $this->has($image->getPath())) {
             $this->remove($image->getPath());
         }
 
         do {
-            $hash = md5(uniqid(mt_rand(), true));
+            $hash = md5(uniqid((string) mt_rand(), true));
             $path = $this->expandPath($hash.'.'.$image->getFile()->guessExtension());
         } while ($this->filesystem->has($path));
 
@@ -46,12 +57,24 @@ class ImageUploader implements ImageUploaderInterface
         );
     }
 
-    public function remove($path)
+    /**
+     * {@inheritdoc}
+     */
+    public function remove(string $path): bool
     {
-        return $this->filesystem->delete($path);
+        if ($this->filesystem->has($path)) {
+            return $this->filesystem->delete($path);
+        }
+
+        return false;
     }
 
-    private function expandPath($path)
+    /**
+     * @param string $path
+     *
+     * @return string
+     */
+    private function expandPath(string $path): string
     {
         return sprintf(
             '%s/%s/%s',
@@ -59,5 +82,15 @@ class ImageUploader implements ImageUploaderInterface
             substr($path, 2, 2),
             substr($path, 4)
         );
+    }
+
+    /**
+     * @param string $path
+     *
+     * @return bool
+     */
+    private function has(string $path): bool
+    {
+        return $this->filesystem->has($path);
     }
 }
